Feature/new xml files

+++ src/Hyperar.HPA.Domain/HattrickEntityBase.cs
@@ -4,6 +4,6 @@
 
     public abstract class HattrickEntityBase : IHattrickEntity
     {
-        public uint HattrickId { get; set; }
+        public long HattrickId { get; set; }
     }
 }
